+++ wls-gui-wahllokalsystem/tests/stores/ergebnismeldungStore.spec.ts
@@ -2,6 +2,7 @@ import type { BezirkUndWahlIDStapelArt } from "@/types/ergebnismeldung/BezirkUnd
 
 import { useCommonTestDataFactory } from "@tests/utils/common/CommonTestDataFactory.ts";
 import { useBegruendungTestDataFactory } from "@tests/utils/ergebnismeldung/begruendungTestDataFactory.ts";
+import { useCommonErgebnismeldungTestDataFactory } from "@tests/utils/ergebnismeldung/commonErgebnismeldungTestDataFactory.ts";
 import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/ergebnisseTestDataFactory.ts";
 import { useUserTestDataFactory } from "@tests/utils/user/UserTestDataFactory.ts";
 import { useWahlTestDataFactory } from "@tests/utils/wahl/WahlTestDataFactory.ts";
@@ -38,16 +39,13 @@ vi.mock(
 
 const { generateRandomString, generateRandomNumber, getRandomItem } =
   useCommonTestDataFactory();
-const {
-  createErgebnis,
-  prepareErgebnis,
-  createErgebnisse,
-  prepareErgebnisse,
-  prepareBezirkUndWahlIDStapelart,
-} = useErgebnisseTestDataFactory();
+const { createErgebnis, prepareErgebnis, createErgebnisse, prepareErgebnisse } =
+  useErgebnisseTestDataFactory();
 const { prepareUser } = useUserTestDataFactory();
 const { prepareWahl } = useWahlTestDataFactory();
 const { createBegruendung } = useBegruendungTestDataFactory();
+const { prepareBezirkUndWahlIDStapelart } =
+  useCommonErgebnismeldungTestDataFactory();
 
 describe("ergebnismeldungStore.ts", () => {
   let unitUnderTest: ReturnType<typeof useErgebnismeldungStore>;
