+++ wls-gui-wahllokalsystem/tests/utils/ergebnismeldung/commonErgebnismeldungTestDataFactory.ts
@@ -1,12 +1,14 @@
 import type { BezirkUndWahlIDStapelartDTO } from "@/api/wls-clients/generated-ergebnismeldung-api";
 import type { BezirkUndWahlIDStapelArt } from "@/types/ergebnismeldung/BezirkUndWahlIDStapelArt.ts";
+import type { Builder } from "@tests/utils/Builder.ts";
 
+import { proxyBuilder } from "@tests/utils/Builder.ts";
 import { useCommonTestDataFactory } from "@tests/utils/common/CommonTestDataFactory.ts";
 
 import { BezirkUndWahlIDStapelartDTOStapelartEnum } from "@/api/wls-clients/generated-ergebnismeldung-api";
 import { StapelArtEnum } from "@/types/ergebnismeldung/StapelArtEnum.ts";
 
-const { generateRandomString } = useCommonTestDataFactory();
+const { generateRandomString, getRandomItem } = useCommonTestDataFactory();
 
 export function useCommonErgebnismeldungTestDataFactory() {
   function createBezirkUndWahlIDStapelartDTO(
@@ -20,17 +22,24 @@ export function useCommonErgebnismeldungTestDataFactory() {
   }
 
   function createBezirkUndWahlIDStapelart(
-    stapelArt: StapelArtEnum
+    stapelArt?: StapelArtEnum
   ): BezirkUndWahlIDStapelArt {
     return {
       wahlID: generateRandomString(10),
       wahlbezirkID: generateRandomString(5),
-      stapelArt: stapelArt,
+      stapelArt: stapelArt ?? getRandomItem(Object.values(StapelArtEnum)),
     };
+  }
+
+  function prepareBezirkUndWahlIDStapelart(): Builder<BezirkUndWahlIDStapelArt> {
+    return proxyBuilder<BezirkUndWahlIDStapelArt>(
+      createBezirkUndWahlIDStapelart()
+    );
   }
 
   return {
     createBezirkUndWahlIDStapelartDTO,
     createBezirkUndWahlIDStapelart,
+    prepareBezirkUndWahlIDStapelart,
   };
 }
